fix: update import paths for index move, refs AC-1202
