+++ apps/web/src/app/admin-console/organizations/members/components/member-dialog/member-dialog.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 
 import { RadioButtonModule } from "@bitwarden/components";
 
-import { SharedOrganizationModule } from "../../../../../organizations/shared";
+import { SharedOrganizationModule } from "../../../shared";
 
 import { MemberDialogComponent } from "./member-dialog.component";
 import { NestedCheckboxComponent } from "./nested-checkbox.component";
