+++ apps/web/src/app/admin-console/organizations/organization.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from "@angular/core";
 
-import { SharedOrganizationModule } from "../../organizations/shared";
-
 import { CoreOrganizationModule } from "./core";
 import { GroupAddEditComponent } from "./manage/group-add-edit.component";
 import { GroupsComponent } from "./manage/groups.component";
 import { OrganizationsRoutingModule } from "./organization-routing.module";
+import { SharedOrganizationModule } from "./shared";
 import { AccessSelectorModule } from "./shared/components/access-selector";
 
 @NgModule({
